+++ src/votekit/graphs/__init__.py
@@ -1,2 +1,6 @@
 from .ballot_graph import BallotGraph  # noqa
-from .pairwise_comparison_graph import PairwiseComparisonGraph  # noqa
+from .pairwise_comparison_graph import (
+    pairwise_dict,  # noqa
+    restrict_pairwise_dict_to_subset,  # noqa
+    PairwiseComparisonGraph,  # noqa
+)