+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Added
+
+## Changed 
+- separated the computation of the pairwise comparison dictionary from the pairwise comparison graph.
+- renamed `dominating_tiers` method of pairwise comparison graph to `get_dominating_tiers`.
+
+## Fixed
 
 ## [3.1.0] - 2025-03-03
 ## Added